Upgrade npm during build
The version of npm included with Alpine 3.13 is currently 6.14.12. This
version seems to have  problem installing node packages from Github. For
example, using a build arg to use a custom version of zwave-js-server
from a github fork + branch fails.

I'm not exactly sure why it fails, but the same exact install command
succeeds using a node:15 docker image, which contains npm 7. Sure
enough, upgrading this build to npm 7 fixes it, for whatever reason.